test: deduplicate validation helper and fixture-ize chart seed setup
- extract shared assert_validation_error_envelope to tests/test_utils.py
- import shared helper in charts/integration test modules and remove duplicates
- refactor _seed_chart_rows_for_get_charts into yield fixture
  - setup seeds chart rows and updates ActiveChartSet before yield
  - teardown restores previous ActiveChartSet and cleans seeded rows
- add tests package marker to avoid mypy duplicate-module resolution
